Samle alt knyttet til korrigering av etterbetaling i en fil og endre filnavn
